+++ types/addr_prefixes.go
@@ -1,6 +1,10 @@
 package types
 
-import sdk "github.com/cosmos/cosmos-sdk/types"
+import (
+	"encoding/binary"
+	"fmt"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
 
 const (
 	// Bech32PrefixAccAddr defines the Bech32 prefix of an account's address
@@ -22,4 +26,14 @@ func init() {
 	config.SetBech32PrefixForAccount(Bech32PrefixAccAddr, Bech32PrefixAccPub)
 	config.SetBech32PrefixForValidator(Bech32PrefixValAddr, Bech32PrefixValPub)
 	config.SetBech32PrefixForConsensusNode(Bech32PrefixConsAddr, Bech32PrefixConsPub)
+	config.SetAddressVerifier(func(bytes []byte) error {
+		n := len(bytes)
+		if n == sdk.AddrLen {
+			return nil
+		}
+		if n <= binary.MaxVarintLen64+1 {
+			return nil
+		}
+		return fmt.Errorf("unexpected address length %d", n)
+	})
 }